+++ CHANGES.rst
@@ -24,6 +24,11 @@ Changes:
   nested under ``weaver.wps_restapi_path`` or ``weaver.wps_restapi_url``, since the OpenAPI `JSON` and `HTML` responses
   are employed for representing supported requests and responses both the `REST` and the `OWS` `WPS` interfaces.
 - Update `Swagger-UI` version for latest rendering fixes of `OpenAPI` definitions.
+- Add multiple missing `OGC API - Processes` conformance references.
+- Modify default query parameter value ``links=true`` for ``/processes`` summary listing to conform with
+  conformance class ``/conf/core/process-summary-links`` as default behavior
+  (relates to `opengeospatial/ogcapi-processes#406 <https://github.com/opengeospatial/ogcapi-processes/pull/406>`_,
+  fixes `crim-ca/weaver#622 <https://github.com/crim-ca/weaver/issues/622>`_).
 - Pin ``gunicorn>=22`` to address CVE-2024-1135.
 
 Fixes: